made InputSet* signals children of Modified
Rationale: When we set an input to a shared pointer, there is no-one to send
the Modified signal. Also, whenever we reset an input (to an Output or shared
pointer), from the targeted process node's point of view, this is a
modification. A (desired) side-effect of this is that you now get your first
Modified signal already on the input setting. If you are really interested in
one of the InputSet* signals, you can add a callback for that. In this case,
your Modified callback will not be called, since there is a more specific one
already.